requirements.txt
accelerate==0.30.1
altair==5.3.0
altair-data-server==0.4.1
altair-viewer==0.4.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.4.2
blinker==1.8.2
cachetools==5.3.3
captum==0.7.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
entrypoints==0.4
Faker==25.3.0
favicon==0.7.0
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
htbuilder==0.6.2
html2image==2.0.4.3
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.2
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.2.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
openai==1.30.5
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
portpicker==1.6.0
prometheus_client==0.20.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.1.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
pydeck==0.9.1
Pygments==2.18.0
pymdown-extensions==10.8.1
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
seaborn==0.13.2
selenium==4.21.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
st-annotated-text==4.0.1
st-theme==1.2.3
streamlit==1.36.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.3
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.12.1
tbb==2021.12.0
tenacity==8.3.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.4
transformers==4.41.2
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
validators==0.28.3
vl-convert-python==1.4.0
watchdog==4.0.1
websocket-client==1.8.0
wsproto==1.2.0
